requirements.txt
asarPy==1.0.1
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
psycopg==3.2.11
psycopg-binary==3.2.11
sqlparse==0.5.3
typing_extensions==4.15.0


